Fix bug while running trace-07-string
Test 07 expect that if the size of the value in queue node is bigger
then the buffer, then the value in queue should be truncated with
a '\0' when copy to buffer.
However, in my original implementation, if size of the node value
is bigger then buffer, then it will copy the value without insert
a '\0' at the end of the buffer.