Implement q_ascend and q_descend
The implementation uses two loop where one loop is inside the other.

The inner loop has two adjacent pointers, if the node of the front
pointer is bigger(or smaller) then the node of the rear pointer,
then the front pointer is deleted.

But iterating the inner loop once cannot delete all elements need to
be deleted, so this method uses the outer loop to run the inner loop
over and over again until all need-to-delete elements are all deleted.